Rename testData folder name to fake
