+++ packages/sync-server/__tests__/fake/fhir.js
@@ -62,4 +62,3 @@ export const fakeResourcesOfFhirServiceRequest = async models => {
 
   return resources;
 };
-
